refactor: rename analysis configuration component
