+++ packages/server-core/src/components/analysis-metadata/types.ts
@@ -5,6 +5,6 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export type AnalysisConfigurationRecalcPayload = {
+export type AnalysisMetadataRecalcPayload = {
     analysisId: string
 };
