+++ packages/server-core/src/components/index.ts
@@ -7,5 +7,5 @@
 
 export * from './constants';
 
-export * from './analysis-configuration';
+export * from './analysis-metadata';
 export * from './registry';
